# src/main/resources/openapi/openApi.yaml
openapi: 3.0.0
info:
  title: Task API
  version: 1.0.0
paths:
  /tasks:
    post:
      summary: Create task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task successfully created
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: Id of created task
        '400':
          description: Error creating task
    get:
      summary: Get list of tasks
      operationId: getAllTasks
      responses:
        '200':
          description: List of all tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDTO'
        '204':
          description: No tasks available
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          description: id of the task
    delete:
      summary: Delete task
      operationId: deleteTask
      responses:
        '200':
          description: Task successfully deleted
        '404':
          description: Task not found
    put:
      summary: Update task status
      operationId: updateTaskStatus
      parameters:
        - schema:
            type: string
          name: status
          in: path
          description: task status
          required: true
      responses:
        '200':
          description: Task status successfully updated
        '400':
          description: Error updating task status
    patch:
      summary: Change task fields
      operationId: changeTaskFields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task fields successfully changed
        '400':
          description: Error changing task fields
components:
  schemas:
    TaskRequest:
      description: Task Request used for creating or updating task
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
        description:
          type: string
        name:
          type: string
    TaskDTO:
      description: Basic Task DTO containing all basic information
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
        - $ref: '#/components/schemas/TaskRequest'
    Status:
      type: string
      description: Task status. Can be CREATE, IN_PROGRESS or DONE
      enum:
        - CREATED
        - IN_PROGRESS
        - DONE